Refactor to a cleaner base+*/nix/windows/unsupported
